requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
pyjwt
passlib
boto3
python-slugify